Création du module OptionsUsed qui englobe toutes les options de commandes (-s, -c, -f, etc).
Petit refactor de CommandFactory (jvais faire plus dans les prochaisn commits)
