+++ src/main/java/com/fges/Main.java
@@ -11,13 +11,8 @@ import java.io.*;
 public class Main {
     public static final ObjectMapper OBJECT_MAPPER = new ObjectMapper();
 
-    public static void main(String[] args) {
-        try {
-            System.exit(exec(args));
-        } catch (IOException ex) {
-            System.err.println("Erreur de fichier: " + ex.getMessage());
-            System.exit(1);
-        }
+    public static void main(String[] args) throws IOException{
+        System.exit(exec(args));
     }
 
     public static int exec(String[] args) throws IOException {
