+++ src/main/java/com/fges/commands/CommandFactory.java
@@ -4,21 +4,18 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fges.groceriesDAO.GroceriesDAO;
 import com.fges.groceriesDAO.GroceriesDAOImplCsv;
 import com.fges.groceriesDAO.GroceriesDAOImplJson;
+import com.fges.modules.OptionsUsed;
 import org.apache.commons.cli.*;
 
 import java.util.List;
 
 public class CommandFactory {
-    public String category;
-    public String fileName;
-    public String format;
+    public OptionsUsed optionsUsed;
     public List<String> command;
 
     public CommandFactory(String[] args) throws ParseException {
         CommandLine cmd = parseArguments(args);
-        this.category = cmd.getOptionValue("c", "default");
-        this.fileName = cmd.getOptionValue("s");
-        this.format = determineFormat(cmd.getOptionValue("f"), this.fileName);
+        this.optionsUsed =  new OptionsUsed(cmd);
         this.command = cmd.getArgList();
     }
 
@@ -31,17 +28,6 @@ public class CommandFactory {
         return parser.parse(cliOptions, args);
     }
 
-    private String determineFormat(String format, String fileName) {
-        if (format == null) {
-            if (fileName != null) {
-                if (fileName.endsWith(".csv")) return "csv";
-                if (fileName.endsWith(".json")) return "json";
-            }
-            return "json"; // Default to json
-        }
-        return format;
-    }
-
     /**
      * Returns the command object based on the command line arguments. It also creates the appropriate DAO I guess.
      *
@@ -50,16 +36,16 @@ public class CommandFactory {
      */
     public Command getCommand(ObjectMapper OBJECT_MAPPER) {
         GroceriesDAO groceriesDAO;
-        switch (this.format) {
-            case "json" -> groceriesDAO = new GroceriesDAOImplJson(this.fileName, OBJECT_MAPPER);
-            case "csv" -> groceriesDAO = new GroceriesDAOImplCsv(this.fileName);
-            default -> throw new IllegalArgumentException("Unknown format: " + this.format);
+        switch (optionsUsed.getFormat()) {
+            case "json" -> groceriesDAO = new GroceriesDAOImplJson(optionsUsed.getFileName(), OBJECT_MAPPER);
+            case "csv" -> groceriesDAO = new GroceriesDAOImplCsv(optionsUsed.getFileName());
+            default -> throw new IllegalArgumentException("Unknown format: " + optionsUsed.getFormat());
         }
         String commandName = this.command.get(0);
         return switch (commandName) {
-            case "add" -> new AddCommand(this.command, groceriesDAO, category);
+            case "add" -> new AddCommand(this.command, groceriesDAO, optionsUsed.getCategory());
             case "list" -> new ListCommand(this.command, groceriesDAO);
-            case "remove" -> new RemoveCommand(this.command, groceriesDAO, category);
+            case "remove" -> new RemoveCommand(this.command, groceriesDAO, optionsUsed.getCategory());
             case "clear" -> new ClearCommand(this.command, groceriesDAO);
             case "info" -> new InfoCommand(this.command);
             default -> throw new IllegalArgumentException("Unknown command: " + commandName);
